+++ .github/workflows/eslint.yml
@@ -16,12 +16,12 @@ jobs:
         node-version: [18.x]
 
     steps:
-      - uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - name: Use Node.js ${{ matrix.node-version }}
-        uses: actions/setup-node@v2
+        uses: actions/setup-node@v3
         with:
           node-version: ${{ matrix.node-version }}
       - name: Install project
         run: npm i
       - name: Lint code
-        run: npm run linter
+        run: npm run linter:check
