+++ frontend/.eslintignore
@@ -1,0 +1,5 @@
+node_modules
+.github
+build
+public
+src/__fixture__
