+++ frontend/package.json
@@ -27,7 +27,8 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "linter": "npx eslint . --fix"
+    "linter": "npx eslint . --fix",
+    "linter:check": "npx eslint ."
   },
   "eslintConfig": {
     "extends": [
