+++ frontend/src/components/Search/SearchContext/index.ts
@@ -1,3 +1,2 @@
-// eslint-disable-next-line no-restricted-exports
 export { default } from './SearchContext';
 export { useSearchContext } from './hooks';
