+++ frontend/src/components/Search/SearchInput/index.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line no-restricted-exports
 export { default } from './SearchInput';
