+++ frontend/src/index.tsx
@@ -2,8 +2,8 @@ import { createRoot } from 'react-dom/client';
 import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-import App from './components/app/App';
-import { setupStore } from './store/store';
+import App from './components/App';
+import { setupStore } from './store';
 
 const container = document.getElementById('root')!;
 const root = createRoot(container);
