+++ frontend/src/store/index.ts
@@ -1,37 +1,25 @@
-import { addFavoriteCountry, setFavoriteCountry } from './reducers/favoriteSlice';
-import { setHistory, addHistory } from './reducers/historySlice';
-import { deleteUser, setUser } from './reducers/userSlice';
-import {
+// countryReducer
+import getAllHistory from './selectors/historySelector';
+import getFavoriteCountries from './selectors/favoriteSelector';
+import getUser from './selectors/userSelector';
+
+export { fetchAllCountry, fetchOneCountry } from './thunks/country';
+export {
   getAllCoutries, getCountriesPaginat, getOneCountry, getMaxPage,
 } from './selectors/countrySelector';
-import getFavoriteCountries from './selectors/favoriteSelector';
-import getAllHistory from './selectors/historySelector';
-import getUser from './selectors/userSelector';
-import type { RootState, AppDispatch } from './store';
-import { fetchAllCountry, fetchOneCountry } from './thunks/country';
-
-// coutryReduser
-export { getCountriesPaginat };
-export { getAllCoutries };
-export { getOneCountry };
-export { fetchAllCountry };
-export { fetchOneCountry };
-export { getMaxPage };
 
 // historyReducer
+export { setHistory, addHistory } from './reducers/historySlice';
 export { getAllHistory };
-export { setHistory };
-export { addHistory };
 
 // favoriteReducer
+export { addFavoriteCountry, setFavoriteCountry } from './reducers/favoriteSlice';
 export { getFavoriteCountries };
-export { setFavoriteCountry };
-export { addFavoriteCountry };
 
 // userReducer
+export { deleteUser, setUser } from './reducers/userSlice';
 export { getUser };
-export { setUser };
-export { deleteUser };
 
-export type { RootState };
-export type { AppDispatch };
+// redux
+export { useAppDispatch, useAppSelector } from './hooks/redux';
+export { type RootState, type AppDispatch, setupStore } from './store';
